+++ R/plotPA.R
@@ -62,13 +62,14 @@ plotPA <- function(map, th, colors = NULL, hr = FALSE, filename = NULL,
 
   my_plot <- rasterVis::gplot(pa, maxpixels = maxpixels)
 
+  # TODO Remove this line if works
   # In some cases rasterVis changes logical values into 0s and 1s
-  my_plot$data$value <- as.logical(my_plot$data$value)
+  #my_plot$data$value <- as.logical(my_plot$data$value)
 
   my_plot <- my_plot +
     geom_tile(aes_(fill = ~value)) +
     scale_fill_manual(values = colors,
-                      breaks = c(1, 0),
+                      breaks = c(TRUE, FALSE),
                       labels = c("Presence", "Absence"),
                       name = "") +
     coord_equal() +